Remove `config.generator.default`
