+++ test/unit/generator/test_pdf.rb
@@ -21,4 +21,21 @@ class Thinreports::Generator::TestPDF < Minitest::Test
   ensure
     PDF::Document.singleton_class.send(:remove_method, :new)
   end
+
+  def test_default_layout
+    layout_filename = layout_file.path
+    report = Thinreports::Report.new layout: layout_filename
+
+    generator = PDF.new report
+    assert_equal layout_filename, generator.default_layout.filename
+  end
+
+  def test_initialize
+    report = Thinreports::Report.new layout: layout_file.path
+    report.start_new_page
+
+    PDF.new(report)
+
+    assert_equal true, report.finalized?
+  end
 end
